requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.5
cryptography==42.0.8
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
gTTS==2.5.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
importlib_resources==6.4.0
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pdfminer==20191125
pdfminer.six==20231228
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
pywin32==306
referencing==0.34.0
requests==2.31.0
resend==2.2.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.31
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
